chore: FeedViewModel 및 sharedRepository간 책임 명확화
